minor fixes, refactoring
