+++ src/components/ProteinTabs.tsx
@@ -1,14 +1,14 @@
 import * as React from "react"
 import { useEffect } from "react"
-import { Link } from "react-router-dom"
 import Box from "@mui/material/Box"
 import Tab from "@mui/material/Tab"
 import Tabs from "@mui/material/Tabs"
-import ProtvistaUniprot from 'protvista-uniprot';
+import ProtvistaUniprot from "protvista-uniprot"
 import styled from "styled-components"
 
 import CopyIcon from "../assets/CopyIcon"
-import ExternalLinkIcon from "../assets/ExternalLinkIcon"
+import { ProteinTabsProps, TabPanelProps } from "../types"
+import Publications from "./Publications"
 
 declare global {
   namespace JSX {
@@ -19,29 +19,6 @@ declare global {
 }
 
 window.customElements.define("protvista-uniprot", ProtvistaUniprot as any)
-
-interface TabPanelProps {
-  children?: React.ReactNode
-  index: number
-  value: number
-}
-
-interface ProteinData {
-  organism: string
-  accession: string
-  uniProtkbId: string
-  description: string
-  gene: string
-  length: number
-  lastUpdated: string
-  mass: number
-  checksum: string
-  sequence: string
-}
-
-interface ProteinTabsProps {
-  data: ProteinData
-}
 
 const TabPanel = (props: TabPanelProps) => {
   const { children, value, index, ...other } = props
@@ -184,130 +161,12 @@ const ProteinTabs: React.FC<ProteinTabsProps> = ({ data }) => {
         </div>
       </TabPanel>
       <TabPanel value={value} index={1}>
-        <div><protvista-uniprot accession={data.accession} /></div>
+        <div>
+          <protvista-uniprot accession={data.accession} />
+        </div>
       </TabPanel>
       <TabPanel value={value} index={2}>
-        <div className="publications">
-          {publications &&
-            publications.map((publication: any) => (
-              <div
-                key={publication.citation.id}
-                className="publication-container"
-              >
-                <h3 className="publication-title">
-                  {publication.citation.title}
-                </h3>
-                <p className="publication-authors">
-                  {publication.citation.authors
-                    ? (publication.citation.authors.length > 15
-                      ? publication.citation.authors
-                          .slice(0, 15)
-                          .map((author: string, index: number) => {
-                            return index ===
-                              publication.citation.authors.slice(0, 15).length -
-                                1 ? (
-                              <span key={index}>
-                                {author}
-                                {"..."}
-                              </span>
-                            ) : (
-                              <span key={index}>
-                                {author}
-                                {", "}
-                              </span>
-                            )
-                          })
-                      : publication.citation.authors.map(
-                          (author: string, index: number) => {
-                            return index ===
-                              publication.citation.authors.length - 1 ? (
-                              <span key={index}>{author}</span>
-                            ) : (
-                              <span key={index}>
-                                {author}
-                                {", "}
-                              </span>
-                            )
-                          },
-                        ))
-                    : ""}
-                </p>
-                <p>
-                  <span className="subtitle-grey-text">{"Categories: "}</span>
-                  {publication.references &&
-                  publication.references[0] &&
-                  publication.references[0].sourceCategories
-                    ? publication.references[0].sourceCategories.join(", ")
-                    : ""}
-                </p>
-                <p>
-                  <span className="subtitle-grey-text">{"Cited for: "}</span>
-                  {publication.references &&
-                  publication.references[0] &&
-                  publication.references[0].referencePositions
-                    ? publication.references[0].referencePositions.join(", ")
-                    : ""}
-                </p>
-                <p>
-                  <span className="subtitle-grey-text">{"Source: "}</span>
-                  <span>
-                    {publication.references &&
-                      publication.references[0] &&
-                      publication.references[0].source.name}
-                  </span>
-                </p>
-
-                {publication.citation.citationCrossReferences &&
-                publication.citation.citationCrossReferences[0] ? (
-                  <div className="publication-links-container">
-                    <Link
-                      className="publication-link"
-                      to={`https://pubmed.ncbi.nlm.nih.gov/${publication.citation.citationCrossReferences[0].id}`}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                    >
-                      {"PubMed "}
-                      {<ExternalLinkIcon />}
-                    </Link>{" "}
-                    <Link
-                      className="publication-link"
-                      to={`https://europepmc.org/article/MED/${publication.citation.citationCrossReferences[0].id}`}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                    >
-                      {"Europe PMC "}
-                      {<ExternalLinkIcon />}
-                    </Link>{" "}
-                    <Link
-                      className={`publication-link ${
-                        !publication.citation.citationCrossReferences[1]?.id &&
-                        "disabled"
-                      }`}
-                      to={
-                        publication.citation?.citationCrossReferences[1]?.id
-                          ? `https://dx.doi.org/${publication.citation.citationCrossReferences[1].id}`
-                          : ""
-                      }
-                      onClick={(event) => {
-                        if (
-                          !publication.citation.citationCrossReferences[1]?.id
-                        ) {
-                          event.preventDefault()
-                        }
-                      }}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                    >
-                      {`${publication.citation.journal} :${publication.citation.firstPage}-${publication.citation.lastPage} (${publication.citation.publicationDate})`}{" "}
-                      {<ExternalLinkIcon />}
-                    </Link>{" "}
-                  </div>
-                ) : (
-                  ""
-                )}
-              </div>
-            ))}
-        </div>
+        <Publications publications={publications} />
       </TabPanel>
     </Box>
   )
@@ -370,72 +229,6 @@ const Wrapper = styled.section`
       justify-content: space-between;
       align-items: flex-start;
       width: 100%;
-    }
-  }
-  .publications {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-    gap: 15px;
-    text-align: left;
-    margin-bottom: 50px;
-
-    .publication-container {
-      display: flex;
-      flex-direction: column;
-      justify-content: space-between;
-      align-items: flex-start;
-      width: 100%;
-      max-width: 800px;
-      padding: 20px 23px;
-      background-color: var(--grey);
-      border-radius: 8px;
-
-      .publication-title {
-        color: var(--dark);
-        font-size: 16px;
-        font-weight: 600;
-        margin-bottom: 12px;
-      }
-      .publication-authors span {
-        color: var(--dark);
-        font-size: 14px;
-        font-weight: 400;
-        text-decoration: underline;
-      }
-      .subtitle-grey-text {
-        color: var(--dark-grey);
-      }
-      p {
-        font-size: 14px;
-      }
-      p:last-of-type {
-        margin-bottom: 23px;
-      }
-      .publication-links-container {
-        display: flex;
-        flex-direction: row;
-        justify-content: flex-start;
-        align-items: center;
-        gap: 10px;
-        .publication-link {
-          border: 1px solid var(--active-blue);
-          border-radius: 8px;
-          padding: 5px 9px;
-          color: var(--active-blue);
-          vertical-align: middle;
-          line-height: 12px;
-        }
-        .publication-link.disabled {
-          border: 1px solid var(--dark-grey-2);
-          color: var(--dark-grey-2);
-          svg path {
-            fill: var(--dark-grey-2);
-          }
-        }
-      }
     }
   }
 `
