+++ src/features/search/searchSlice.ts
@@ -5,11 +5,7 @@ import {
 } from "firebase/auth"
 
 import { auth } from "../../firebase"
-import { Filters, Filter, SearchState } from "../../types"
-
-
-
-
+import { Filter, Filters, SearchState } from "../../types"
 
 const initialState: SearchState = {
   currentUser: null,
@@ -86,39 +82,45 @@ export const searchSlice = createSlice({
     setResults: (state, action) => {
       state.results = action.payload
     },
-setFilters: (state, action) => {
-  if (!action.payload) {
-    state.selectedFilters = null;
-    return;
-  }
+    setFilters: (state, action) => {
+      if (!action.payload) {
+        state.selectedFilters = null
 
-  const newFilters: Filters = action.payload.reduce((accumulator: Filters, filter: Filter) => {
-    const existingFilterIndex = accumulator.findIndex((f) => f.name === filter.name);
-
-    if (existingFilterIndex !== -1) {
-      const existingFilter = accumulator[existingFilterIndex];
-
-      if (filter.name === "length") {
-        accumulator[existingFilterIndex] = {
-          ...existingFilter,
-          minLength: filter.minLength,
-          maxLength: filter.maxLength,
-        };
-      } else {
-        accumulator[existingFilterIndex] = {
-          ...existingFilter,
-          value: filter.value,
-        };
+        return
       }
-    } else {
-      accumulator.push(filter);
-    }
 
-    return accumulator;
-  }, []);
+      const newFilters: Filters = action.payload.reduce(
+        (accumulator: Filters, filter: Filter) => {
+          const existingFilterIndex = accumulator.findIndex(
+            (f) => f.name === filter.name,
+          )
 
-  state.selectedFilters = newFilters;
-},
+          if (existingFilterIndex !== -1) {
+            const existingFilter = accumulator[existingFilterIndex]
+
+            if (filter.name === "length") {
+              accumulator[existingFilterIndex] = {
+                ...existingFilter,
+                minLength: filter.minLength,
+                maxLength: filter.maxLength,
+              }
+            } else {
+              accumulator[existingFilterIndex] = {
+                ...existingFilter,
+                value: filter.value,
+              }
+            }
+          } else {
+            accumulator.push(filter)
+          }
+
+          return accumulator
+        },
+        [],
+      )
+
+      state.selectedFilters = newFilters
+    },
     setIsFiltersModalOpen: (state, action) => {
       state.isFiltersModalOpen = action.payload
     },
